requirements.txt
streamlit>=1.29.0
docling>=1.0.0
pinecone>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
sentence-transformers>=2.2.0
tiktoken>=0.5.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
requests>=2.31.0